update robot and image_process.py

+++ robot.py
@@ -6,8 +6,8 @@ import argparse
 import cv2
 import time
 import atexit
-from multiprocessing import Process, Manager
-import image_processor
+from multiprocessing import Process, Manager, Queue
+from image_processor import ImageCalculater
 import detection
 from tts import loudspeaker
 
@@ -22,29 +22,31 @@ class Robot(object):
     self.frame_height = 240
     self.camera = None
     self.debug = True
-    self.plist = []
-    self.image_processor = image_processor.ImageCalculater(self.frame_width, self.frame_height)
-    self.speaker = loudspeaker()
+    self.image_processor = ImageCalculater(self.frame_width, self.frame_height)
+
+    # self.speaker = loudspeaker()
     # speaker.raise_alert()
 
   def release_robot_resources(self):
     print "You are now leaving the Python sector."
     self.camera.release()
-    cv2.destroyAllWindows()
-    gpio.release_robot_gpio()
-    self.speaker.release_tts()
+    # gpio.release_robot_gpio()
+    # self.speaker.release_tts()
 
-    for p in self.plist:
-      p.join()
+    if self.debug:
+      cv2.destroyAllWindows()
 
-  def init_devices(self):
+  def init_resources(self):
     self.camera = cv2.VideoCapture(self.video)
     self.image_processor.camera = self.camera
 
-    gpio.init_robot_gpio()
-    gpio.set_speed(ctypes.c_float(0.8))
+    # gpio.init_robot_gpio()
+    # gpio.set_speed(ctypes.c_float(0.8))
 
     atexit.register(self.release_robot_resources)
+    if self.debug:
+      cv2.namedWindow("origin_frame")
+      # cv2.namedWindow("frame")
 
   def search_object_camshift(self):
     frame, contour = self.image_processor.search_by_optical_flow()
@@ -78,49 +80,60 @@ class Robot(object):
 
     return False
 
-  def search_and_track_object(self, tacking_data):
-    print('start to search_and_track_object')
-    self.image_processor.search_by_optical_flow()
-    while True:
-      if self.image_processor.tracking_by_optical_flow() is 'lost':
-        print('lost object')
-        self.image_processor.search_by_optical_flow()
-      else:
-        tacking_data = self.image_processor.tracking_data_list
-        print("search_and_track_object list count %d"%(len(self.tacking_data)))
+def start_tracking(q):
+  robot = Robot()
+  robot.init_resources()
 
-  def checking_object_falling(self, tacking_data):
-    while True:
-      if len(tacking_data) > 0:
-        print("checking_object_falling list count %d"%(len(tacking_data)))
-      if detection.check_object_status(tacking_data) is False:
-        self.speaker.raise_alert()
+  image_processor = ImageCalculater(robot.frame_width, robot.frame_height)
+  image_processor.camera = robot.camera
+  image_processor.search_by_optical_flow()
 
-  def start(self):
-    # p = Process(target = download, args = (node, exec_cmd));
-    tacking_data = []
+  while True:
+    ret = image_processor.tracking_by_optical_flow()
+    if ret is True:
+      q.put(image_processor.tracking_data_list)
+    elif ret is 'lost':
+      print('lost object...')
+      image_processor.search_by_optical_flow()
+      # tracking_data_list = robot.image_processor.tracking_data_list
+      # print("search_and_track_object list count %d"%(len(tracking_data_list)))
 
-    tracking_worker = Process(target = self.search_and_track_object, args=(tacking_data))
-    tracking_worker.daemon = True
+def start_detecting(q):
+  print('start detection')
+  while True:
+    data_list = q.get()
+    # print(data_list)
+    if len(data_list) > 5:
+      if detection.check_object_status(data_list) is False:
+        print('someone is falling down...')
+        # break
+        # loudspeaker.raise_alert()
 
-    tracking_worker.start()
-
-    detect_falling_worker = Process(target = self.checking_object_falling, args=(tacking_data))
-    detect_falling_worker.daemon = True
-    detect_falling_worker.start()
-
-    self.plist.append(tracking_worker)
-    self.plist.append(detect_falling_worker)
-
-    for p in self.plist:
-      p.join()
+def start_moving(q):
+  print('start moving')
+  while True:
+    if len(q.get()) > 0:
+      print("start_detection list count %d"%(len(q.get())))
 
 def main():
-  robot = Robot('./videos/robot.mp4')
-  # robot = Robot()
-  robot.init_devices()
 
-  robot.start()
+  q = Queue()
+
+  tracking_worker = Process(target = start_tracking, args=(q,))
+  tracking_worker.daemon = True
+  tracking_worker.start()
+
+  detect_falling_worker = Process(target = start_detecting, args=(q,))
+  detect_falling_worker.daemon = True
+  detect_falling_worker.start()
+
+  moving_worker = Process(target = start_moving, args=(q,))
+  moving_worker.daemon = True
+  # moving_worker.start()
+
+  detect_falling_worker.join()
+  tracking_worker.join()
+  moving_worker.join()
 
 if __name__ == "__main__":
   main()
